refactor(database, controller, build.gradle): :recycle: refatorando para sem requisicoes http, requisicoes pela view
