+++ EmprestAI/build.gradle
@@ -11,23 +11,10 @@ repositories {
 }
 
 dependencies {
-    implementation 'mysql:mysql-connector-java:8.0.33'  // Conector MySQL para JDBC
-
-    // Vamos implementar dependencia pra JSON??
+    implementation "mysql:mysql-connector-java:8.0.33"  // Conector MySQL para JDBC
 
     testImplementation platform('org.junit:junit-bom:5.10.0')
     testImplementation 'org.junit.jupiter:junit-jupiter'
-
-    // (Sugestão) Biblioteca para log de ações no gerenciador
-    implementation 'org.slf4j:slf4j-api:2.0.12'
-    implementation 'ch.qos.logback:logback-classic:1.5.3'
-
-    //sparkjava para configurar o servidor
-    implementation 'io.javalin:javalin:5.6.3'
-    implementation 'com.google.code.gson:gson:2.10.1'
-    implementation 'mysql:mysql-connector-java:8.0.33'
-    implementation 'org.slf4j:slf4j-simple:2.0.7'
-    implementation 'com.fasterxml.jackson.core:jackson-databind:2.15.2'
 }
 
 test {
