--- EmprestAI/src/main/java/br/com/emprestai/dataBaseConfig/jsonUtil/JsonUtil.java
@@ -1,16 +1,0 @@
-package br.com.emprestai.dataBaseConfig.jsonUtil;
-
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-
-public class JsonUtil {
-    private static final Gson GSON = new GsonBuilder().setPrettyPrinting().create();
-
-    public static String toJson(Object object) {
-        return GSON.toJson(object);
-    }
-
-    public static <T> T fromJson(String json, Class<T> clazz) {
-        return GSON.fromJson(json, clazz);
-    }
-}
