+++ EmprestAI/src/main/java/br/com/emprestai/database/DatabaseConnection.java
@@ -1,17 +1,17 @@
-package br.com.emprestai.dataBaseConfig;
+package br.com.emprestai.database;
 
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.SQLException;
 
-public class DatabaseConfig {
-    private static final String URL = "jdbc:mysql://mysql-emprestai-emprestai.l.aivencloud.com:15334/emprestai_db?useSSL=false&serverTimezone=UTC";
+public class DatabaseConnection {
+    private static final String URL = "";
 
 
     public static Connection getConnection() throws SQLException {
         try {
             Class.forName("com.mysql.cj.jdbc.Driver");
-            return DriverManager.getConnection(URL, USER, PASSWORD);
+            return null;
+            //return DriverManager.getConnection(URL, USER, PASSWORD);
         } catch (ClassNotFoundException e) {
             throw new SQLException("Driver MySQL não encontrado", e);
         }
