+++ EmprestAI/src/main/java/br/com/emprestai/database/exception/ApiException.java
@@ -1,4 +1,4 @@
-package br.com.emprestai.dataBaseConfig.exception;
+package br.com.emprestai.database.exception;
 
 public class ApiException extends RuntimeException {
     private final int statusCode;
